Adjusted QR interface to generate_wallet
