+++ pyproject.toml
@@ -10,7 +10,6 @@ click = "^7.1.1"
 eth-keyfile = "^0.5.1"
 web3 = "^5.7.0"
 qrcode = {extras = ["pil"], version = "^6.1"}
-xxhash = "^1.4.3"
 python-stdnum = "^1.13"
 structlog = "^20.1.0"
 colorama = "^0.4.3"
